+++ CHANGELOG.md
@@ -31,6 +31,7 @@ This new weather input acquires current and future weather conditions from openw
  - Fix execution of Input stop_input()
  - Fix Input Pre-Outputs not turning on
  - Fix Output not activating for Camera
+ - Fix PWM trigger and Duration Method ([#937](https://github.com/kizniche/mycodo/issues/937))
 
 ### Features